Stat screen fully funcional, code for stat collecting added, upgraded whole UI to also colors for each era, did some last changes and added documentation for the app

+++ app/src/main/java/com/example/eraclicker/model/Upgrade.kt
@@ -3,13 +3,13 @@ package com.example.eraclicker.model
 data class Upgrade(
     val id: Int,
     val name: String,
-    val baseCost: Int,
-    val bonus: Int,
+    val baseCost: Long,
+    val bonus: Long,
     val type: UpgradeType,
     val era: Int,
     val imageRes: Int,
     val level: Int = 0
 ) {
-    val cost: Int
+    val cost: Long
         get() = baseCost * (level + 1)
 }
